--- a/src/NServiceBus.Persistence.AzureTable/NServiceBus.Persistence.AzureTable.csproj
+++ b/src/NServiceBus.Persistence.AzureTable/NServiceBus.Persistence.AzureTable.csproj
@@ -7,6 +7,7 @@
     <Description>NServiceBus Azure Table Persistence</Description>
     <RootNamespace>NServiceBus.Persistence.AzureTable</RootNamespace>
     <LangVersion>10.0</LangVersion>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
   <ItemGroup>