https://github.com/Particular/NServiceBus.Persistence.CosmosDB/pull/300

|                             Method |     Mean |   Error |  StdDev | Ratio |  Gen 0 | Allocated |
|----------------------------------- |---------:|--------:|--------:|------:|-------:|----------:|
|                   Original| 729.7 ns | 2.78 ns | 2.60 ns |  1.00 | 0.0458 |     392 B |
|                        Netframework Optimized | 720.4 ns | 6.16 ns | 5.76 ns |  0.99 | 0.0260 |     224 B |
| Netcore Optimized| 624.1 ns | 3.22 ns | 2.85 ns |  0.86 | 0.0146 |     128 B |

This is something for the master branch I wanted to avoid pushing into the current release just yet. 

The benchmark https://github.com/danielmarbach/MicroBenchmarks/commit/8c5fd4b1671f6aad15f74a238937e9234ebd90f9